requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain & LangGraph (2025 Standards)
langchain==0.2.16
langchain-openai==0.1.22
langchain-anthropic==0.1.23
langgraph==0.2.16
langchain-core==0.2.39

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tiktoken==0.7.0  # Token counting for 2025 optimization

# Caching
redis==5.0.1

# HTTP & CORS
httpx==0.26.0
